fix: manually merge two article generators to one file
